[Modify] 단순 함수 위치 조정
